Makefile: clear out LDFLAGS, add new (cleared) LIBS variable and pass down as QMAKE_LIBS.
The previous behavior was leading to compile failures on arm64 machines
(albeit only old distro versions, apparently), because QMAKE_LFLAGS is
being put on the command line *before* the actual object files to be
included/linked, which is an error for system libraries.

Builds will still fail after this commit, but that's okay.