src/sshmasterconnection.{cpp,h}: skip checkLogin() if hostname starts with "!".
Some special ssh proxies will not allow arbitrary commands. checkLogin()
will break these sessions because it tries to run the echo command on
the proxy.

By specifying a "!" as the first character of the (proxy) hostname you
can instruct x2goclient ot skip the checkLogin() call altogether. Note
that this will break proxies that require you to change you password or
some other type of interaction.

As this is added to SshMasterConnection it is also valid to specify that
for the server hostname although this is not very useful.