src/compat.h: add more compat code for SplitBehavior.
This whole thing is getting really unwieldy, but there's no easy to do
it otherwise.

We can't create implicit conversions for enums and macro magic doesn't
help either, so subclassing QString, adding compat wrappers and using a
macro to change all occurrences of QString to QStringCompatWrapper seems
to be the only way to pull this off.

In a rather ugly way.